+++ test/filter_func_tests.cc
@@ -7,7 +7,7 @@ TEST_CASE("Test window functions", "[WindowFuncs][Helpers]") {
   SECTION("Test Hann window") {
     unsigned int window_length = 23;
     auto hann_window =
-        Dispatcher<std::vector<double>, unsigned int>::instance()->dispatch(
+        Dispatcher<std::vector<double>, double>::instance()->dispatch(
             "HannWindow", window_length);
 
     // This is really is simply function using std::cos function evaluated at
